pyproject.toml
[tool.poetry]
name = "wordtoexcel"
version = "0.1.0"
description = "test"
authors = ["Leon Lee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.11"
pandas = "1.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
